🔐 Добавлена поддержка назначения администратора
- В модель User возвращено поле `is_admin: bool` (по умолчанию False)
- Сгенерирована и применена миграция Alembic для добавления поля `is_admin`
- Назначен конкретный пользователь админом через SQL (обновление поля is_admin = TRUE)
- Пользователи с is_admin = TRUE теперь могут обращаться к маршрутам admin_service
- Проверено: верификация email работает, авторизация через токен — тоже

✅ Теперь можно управлять пользователями с правами администратора
