+++ auth_service/alembic/versions/786527123c15_baseline.py
@@ -1,0 +1,28 @@
+"""baseline
+
+Revision ID: 786527123c15
+Revises: 
+Create Date: 2025-05-05 10:08:44.841519
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '786527123c15'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    pass
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    pass
