+++ auth_service/alembic/versions/8d682b4ef273_remove_verification_code_from_users.py
@@ -1,8 +1,8 @@
-"""Add is_active field to User
+"""remove verification_code from users
 
-Revision ID: 569f0de56a89
-Revises: f9189a322cf9
-Create Date: 2025-04-07 11:29:39.381863
+Revision ID: 8d682b4ef273
+Revises: 786527123c15
+Create Date: 2025-05-05 10:23:21.301660
 
 """
 from typing import Sequence, Union
@@ -12,8 +12,8 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision: str = '569f0de56a89'
-down_revision: Union[str, None] = 'f9189a322cf9'
+revision: str = '8d682b4ef273'
+down_revision: Union[str, None] = '786527123c15'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
@@ -21,12 +21,14 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     """Upgrade schema."""
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('is_active', sa.Boolean(), nullable=False))
+    op.add_column('users', sa.Column('is_admin', sa.Boolean(), nullable=False))
+    op.drop_column('users', 'verification_code')
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     """Downgrade schema."""
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'is_active')
+    op.add_column('users', sa.Column('verification_code', sa.VARCHAR(length=6), autoincrement=False, nullable=True))
+    op.drop_column('users', 'is_admin')
     # ### end Alembic commands ###
