--- auth_service/alembic/versions/a21c3368e2ff_force_add_verification_fields.py
@@ -1,34 +1,0 @@
-"""force add verification fields
-
-Revision ID: a21c3368e2ff
-Revises: 569f0de56a89
-Create Date: 2025-05-04 19:39:49.206273
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'a21c3368e2ff'
-down_revision: Union[str, None] = '569f0de56a89'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('is_verified', sa.Boolean(), nullable=False))
-    op.add_column('users', sa.Column('verification_code', sa.String(length=6), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'verification_code')
-    op.drop_column('users', 'is_verified')
-    # ### end Alembic commands ###
