+++ auth_service/models.py
@@ -15,8 +15,10 @@ class User(Base):
     is_active: Mapped[bool] = mapped_column(Boolean, default=False)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=func.now())
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=func.now(), onupdate=func.now())
-    is_verified: Mapped[bool] = mapped_column(Boolean, default=False)
     verification_code: Mapped[str] = mapped_column(String(6), nullable=True)
+    is_verified: Mapped[bool] = mapped_column(Boolean, default=False)
+    is_admin: Mapped[bool] = mapped_column(Boolean, default=False)
+
 
     # # Связь с постами
     # posts: Mapped[list['Post']] = relationship('Post', back_populates='author')
